move samples to sample package (#104)
