+++ src/handlers/user_handler_test.go
@@ -14,7 +14,7 @@ import (
 	kts_errors "github.com/ELITE-Kinoticketsystem/Backend-KTS/src/errors"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/mocks"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/models"
-	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/utils"
+	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/samples"
 )
 
 func TestRegisterUser(t *testing.T) {
@@ -27,21 +27,21 @@ func TestRegisterUser(t *testing.T) {
 	}{
 		{
 			name: "Success",
-			body: utils.GetSampleRegistrationData(),
+			body: samples.GetSampleRegistrationData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, registrationData interface{}) {
 				mockController.EXPECT().RegisterUser(registrationData).Return(
 					&models.LoginResponse{
-						User: utils.GetSampleUser(),
+						User: samples.GetSampleUser(),
 						/* Token */
 						/* RefreshToken */
 					}, nil)
 			},
-			expectedResponseBody: utils.GetSampleUser(),
+			expectedResponseBody: samples.GetSampleUser(),
 			expectedStatus:       http.StatusCreated,
 		},
 		{
 			name: "Internal Error",
-			body: utils.GetSampleRegistrationData(),
+			body: samples.GetSampleRegistrationData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, registrationData interface{}) {
 				mockController.EXPECT().RegisterUser(registrationData).Return(nil, kts_errors.KTS_INTERNAL_ERROR)
 			},
@@ -52,7 +52,7 @@ func TestRegisterUser(t *testing.T) {
 		},
 		{
 			name: "Email exists",
-			body: utils.GetSampleRegistrationData(),
+			body: samples.GetSampleRegistrationData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, registrationData interface{}) {
 				mockController.EXPECT().RegisterUser(registrationData).Return(nil, kts_errors.KTS_EMAIL_EXISTS)
 			},
@@ -63,7 +63,7 @@ func TestRegisterUser(t *testing.T) {
 		},
 		{
 			name: "Upstream Error",
-			body: utils.GetSampleRegistrationData(),
+			body: samples.GetSampleRegistrationData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, registrationData interface{}) {
 				mockController.EXPECT().RegisterUser(registrationData).Return(nil, kts_errors.KTS_UPSTREAM_ERROR)
 			},
@@ -161,7 +161,7 @@ func TestLoginUser(t *testing.T) {
 		},
 		{
 			name: "User not found",
-			body: utils.GetSampleLoginData(),
+			body: samples.GetSampleLoginData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, loginData models.LoginRequest) {
 				mockController.EXPECT().LoginUser(loginData).Return(nil, kts_errors.KTS_USER_NOT_FOUND)
 			},
@@ -172,7 +172,7 @@ func TestLoginUser(t *testing.T) {
 		},
 		{
 			name: "Internal error",
-			body: utils.GetSampleLoginData(),
+			body: samples.GetSampleLoginData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, loginData models.LoginRequest) {
 				mockController.EXPECT().LoginUser(loginData).Return(nil, kts_errors.KTS_INTERNAL_ERROR)
 			},
@@ -183,7 +183,7 @@ func TestLoginUser(t *testing.T) {
 		},
 		{
 			name: "Incorrect password",
-			body: utils.GetSampleLoginData(),
+			body: samples.GetSampleLoginData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, loginData models.LoginRequest) {
 				mockController.EXPECT().LoginUser(loginData).Return(nil, kts_errors.KTS_CREDENTIALS_INVALID)
 			},
@@ -194,9 +194,9 @@ func TestLoginUser(t *testing.T) {
 		},
 		{
 			name: "Success",
-			body: utils.GetSampleLoginData(),
+			body: samples.GetSampleLoginData(),
 			setExpectations: func(mockController *mocks.MockUserControllerI, loginData models.LoginRequest) {
-				user := utils.GetSampleUser()
+				user := samples.GetSampleUser()
 				mockController.EXPECT().LoginUser(loginData).Return(
 					&models.LoginResponse{
 						User: user,
@@ -204,7 +204,7 @@ func TestLoginUser(t *testing.T) {
 						/* RefreshToken */
 					}, nil)
 			},
-			expectedResponseBody: utils.GetSampleUser(),
+			expectedResponseBody: samples.GetSampleUser(),
 			expectedStatus:       http.StatusOK,
 		},
 	}
