+++ src/repositories/price_category_repository_test.go
@@ -9,13 +9,14 @@ import (
 	kts_errors "github.com/ELITE-Kinoticketsystem/Backend-KTS/src/errors"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/managers"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/models"
+	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/samples"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/utils"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestGetPriceCategories(t *testing.T) {
-	samplePriceCategories := utils.GetSamplePriceCategories()
+	samplePriceCategories := samples.GetSamplePriceCategories()
 
 	query := "SELECT price_categories.id AS \"price_categories.id\", price_categories.category_name AS \"price_categories.category_name\", price_categories.price AS \"price_categories.price\" FROM `KinoTicketSystem`.price_categories;"
 
@@ -99,7 +100,7 @@ func TestGetPriceCategories(t *testing.T) {
 }
 
 func TestGetPriceCategoryById(t *testing.T) {
-	samplePriceCategory := utils.GetSamplePriceCategory()
+	samplePriceCategory := samples.GetSamplePriceCategory()
 
 	priceCategoryId := samplePriceCategory.ID
 
@@ -183,7 +184,7 @@ func TestGetPriceCategoryById(t *testing.T) {
 }
 
 func TestCreatePriceCategory(t *testing.T) {
-	samplePriceCategory := utils.GetSamplePriceCategory()
+	samplePriceCategory := samples.GetSamplePriceCategory()
 
 	query := "INSERT INTO `KinoTicketSystem`.price_categories (id, category_name, price) VALUES (?, ?, ?);"
 
@@ -266,7 +267,7 @@ func TestCreatePriceCategory(t *testing.T) {
 }
 
 func TestUpdatePriceCategory(t *testing.T) {
-	samplePriceCategory := utils.GetSamplePriceCategory()
+	samplePriceCategory := samples.GetSamplePriceCategory()
 
 	query := "UPDATE `KinoTicketSystem`.price_categories SET id = ?, category_name = ?, price = ? WHERE price_categories.id = ?;"
 
@@ -287,7 +288,7 @@ func TestUpdatePriceCategory(t *testing.T) {
 		{
 			name: "Error while updating PriceCategory",
 			setExpectations: func(mock sqlmock.Sqlmock, priceCategory *model.PriceCategories) {
-				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(),  priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnError(sqlmock.ErrCancelled)
+				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(), priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnError(sqlmock.ErrCancelled)
 			},
 			expectedPriceCategoryID: false,
 			expectedError:           kts_errors.KTS_INTERNAL_ERROR,
@@ -295,7 +296,7 @@ func TestUpdatePriceCategory(t *testing.T) {
 		{
 			name: "Error while converting rows affected",
 			setExpectations: func(mock sqlmock.Sqlmock, priceCategory *model.PriceCategories) {
-				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(),  priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnResult(
+				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(), priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnResult(
 					sqlmock.NewErrorResult(errors.New("rows affected conversion did not work")),
 				)
 			},
@@ -305,7 +306,7 @@ func TestUpdatePriceCategory(t *testing.T) {
 		{
 			name: "PriceCategory not found",
 			setExpectations: func(mock sqlmock.Sqlmock, priceCategory *model.PriceCategories) {
-				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(),  priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnResult(sqlmock.NewResult(1, 0))
+				mock.ExpectExec(query).WithArgs(sqlmock.AnyArg(), priceCategory.CategoryName, priceCategory.Price, utils.EqUUID(priceCategory.ID)).WillReturnResult(sqlmock.NewResult(1, 0))
 			},
 			expectedPriceCategoryID: false,
 			expectedError:           kts_errors.KTS_NOT_FOUND,
@@ -348,7 +349,7 @@ func TestUpdatePriceCategory(t *testing.T) {
 }
 
 func TestDeletePriceCategory(t *testing.T) {
-	samplePriceCategory := utils.GetSamplePriceCategory()
+	samplePriceCategory := samples.GetSamplePriceCategory()
 
 	priceCategoryId := samplePriceCategory.ID
