+++ src/samples/pricecategories_samples.go
@@ -3,6 +3,7 @@ package samples
 import (
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/.gen/KinoTicketSystem/model"
 	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/utils"
+	"github.com/google/uuid"
 )
 
 func GetPriceCategories() *[]model.PriceCategories {
@@ -23,4 +24,39 @@ func GetPriceCategories() *[]model.PriceCategories {
 			Price:        5,
 		},
 	}
+}
+
+func GetSamplePriceCategories() *[]model.PriceCategories {
+	priceCategories := []model.PriceCategories{}
+
+	uuid1 := uuid.New()
+	uuid2 := uuid.New()
+
+	priceCategories = append(priceCategories, model.PriceCategories{
+		ID:           &uuid1,
+		CategoryName: "StudentDiscount",
+		Price:        1000,
+	})
+
+	priceCategories = append(priceCategories, model.PriceCategories{
+		ID:           &uuid2,
+		CategoryName: "regular_price",
+		Price:        500,
+	})
+
+	return &priceCategories
+}
+
+func GetSamplePriceCategory() *model.PriceCategories {
+	priceCategory := model.PriceCategories{}
+
+	uuid1 := uuid.MustParse("6ba7b820-9dad-11d1-80b4-00c04fd430c4")
+
+	priceCategory = model.PriceCategories{
+		ID:           &uuid1,
+		CategoryName: "StudentDiscount",
+		Price:        1000,
+	}
+
+	return &priceCategory
 }
