+++ src/samples/user_samples.go
@@ -1,0 +1,39 @@
+package samples
+
+import (
+	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/.gen/KinoTicketSystem/model"
+	"github.com/ELITE-Kinoticketsystem/Backend-KTS/src/models"
+	"github.com/google/uuid"
+)
+
+func GetSampleRegistrationData() models.RegistrationRequest {
+	return models.RegistrationRequest{
+		Username:  "Collinho el niño",
+		Email:     "collin.forslund@gmail.com",
+		Password:  "Passwort",
+		FirstName: "Collin",
+		LastName:  "Forslund",
+	}
+}
+
+func GetSampleLoginData() models.LoginRequest {
+	return models.LoginRequest{
+		Username: "Collinho el niño",
+		Password: "Passwort",
+	}
+}
+
+func GetSampleUser() model.Users {
+	id, _ := uuid.Parse("47CF752501DF45B7A3A9D3CB25AE939F")
+	username := "Collinho el niño"
+	firstname := "Collin"
+	lastname := "Forslund"
+	return model.Users{
+		ID:        &id,
+		Username:  &username,
+		Email:     "collin.forslund@gmail.com",
+		Password:  "$2a$10$vxXPPpLp5baQ7mzS1pNSEuk6ZW3mbx1Ej7u0tJnF5wferEFqT.qlK",
+		Firstname: &firstname,
+		Lastname:  &lastname,
+	}
+}
